+++ src/main/java/com/box/sdk/BoxFileRequest.java
@@ -42,7 +42,7 @@ public class BoxFileRequest extends BoxResource {
         BoxAPIRequest request = new BoxAPIRequest(this.getAPI(), url, "GET");
         BoxJSONResponse response = (BoxJSONResponse) request.send();
         JsonObject responseJSON = JsonObject.readFrom(response.getJSON());
-        return new Info(responseJSON);
+        return new Info(responseJSON, this.getAPI().getBaseAppUrl());
     }
 
     /**
@@ -62,7 +62,7 @@ public class BoxFileRequest extends BoxResource {
         request.setBody(body.toString());
         BoxJSONResponse response = (BoxJSONResponse) request.send();
         JsonObject jsonObject = JsonObject.readFrom(response.getJSON());
-        return new Info(jsonObject);
+        return new Info(jsonObject, this.getAPI().getBaseAppUrl());
     }
 
     /**
@@ -90,7 +90,7 @@ public class BoxFileRequest extends BoxResource {
         BoxJSONResponse response = (BoxJSONResponse) request.send();
         JsonObject jsonObject = JsonObject.readFrom(response.getJSON());
         info.update(jsonObject);
-        return new Info(jsonObject);
+        return new Info(jsonObject, this.getAPI().getBaseAppUrl());
     }
 
     /**
@@ -145,6 +145,8 @@ public class BoxFileRequest extends BoxResource {
         private Date updatedAt;
         private BoxUser.Info updatedBy;
         private URL url;
+        private String path;
+        private String baseUrl;
 
         /**
          * Constructs an empty Info object.
@@ -167,8 +169,14 @@ public class BoxFileRequest extends BoxResource {
          *
          * @param jsonObject the parsed JSON object.
          */
-        Info(JsonObject jsonObject) {
+        Info(JsonObject jsonObject, String fileRequestBaseUrl) {
             super(jsonObject);
+            try {
+                this.baseUrl = fileRequestBaseUrl;
+                this.url = new URL(this.baseUrl + this.path);
+            } catch (MalformedURLException e) {
+                throw new BoxAPIException("Couldn't construct url for file request", e);
+            }
         }
 
         @Override
@@ -360,10 +368,40 @@ public class BoxFileRequest extends BoxResource {
         /**
          * Gets the URL can be shared with users to let them upload files to the associated folder.
          *
-         * @return the date at which this task is due.
+         * @return the URL for files upload.
          */
         public URL getUrl() {
             return this.url;
+        }
+
+        /**
+         * Gets the base URL for the upload files link.
+         *
+         * @return the base url including protocol and hostname.
+         */
+        public String getBaseUrl() {
+            return this.baseUrl;
+        }
+
+        /**
+         * Gets the URL containing only the path (e.g. "/f/123456789") shared with users to let
+         * them upload files to the associated folder.
+         *
+         * @return the path of the URL for files upload.
+         */
+        public String getPath() {
+            return this.path;
+        }
+
+        /**
+         * Sets the base URL for the upload files link. Can throw an exception if format of the URL is invalid.
+         *
+         * @param baseUrl the base url including protocol and hostname.
+         * @throws MalformedURLException when baseUrl format is invalid.
+         */
+        public void setBaseUrl(String baseUrl) throws MalformedURLException {
+            this.baseUrl = baseUrl;
+            this.url = new URL(this.baseUrl + this.path);
         }
 
         @Override
@@ -409,12 +447,7 @@ public class BoxFileRequest extends BoxResource {
                     BoxUser user = new BoxUser(getAPI(), userID);
                     this.createdBy = user.new Info(userJSON);
                 } else if (memberName.equals("url")) {
-                    try {
-                        String urlString = value.asString();
-                        this.url = new URL(urlString);
-                    } catch (MalformedURLException e) {
-                        throw new BoxAPIException("Couldn't parse url for file request", e);
-                    }
+                    this.path = value.asString();
                 }
             } catch (Exception e) {
                 throw new BoxDeserializationException(memberName, value.toString(), e);